requirements.txt

python-dotenv
PyGithub
requests
pydantic>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.7.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiohttp>=3.8.0
prometheus-client>=0.12.0